chore: Move App to components folder
